QA Fix: Implement Data Reliability & Integrity Patches (ACID, Race Conditions, Numeric Precision, Backups)

+++ backend/models/Bin.js
@@ -7,12 +7,12 @@ const Bin = sequelize.define("Bin", {
   rack: { type: DataTypes.STRING },
   shelf: { type: DataTypes.STRING },
   bin_code: { type: DataTypes.STRING, allowNull: false },
-  temperature: { type: DataTypes.FLOAT },
-  humidity: { type: DataTypes.FLOAT },
+  temperature: { type: DataTypes.DECIMAL(15, 2) },
+  humidity: { type: DataTypes.DECIMAL(15, 2) },
   lastAccessed: { type: DataTypes.DATE },
   accessCount: { type: DataTypes.INTEGER, defaultValue: 0 },
   itemCount: { type: DataTypes.INTEGER, defaultValue: 0 },
-  capacityUsed: { type: DataTypes.FLOAT, defaultValue: 0 },
+  capacityUsed: { type: DataTypes.DECIMAL(15, 2), defaultValue: 0 },
   itemsStored: { type: DataTypes.JSON, defaultValue: [] }
 });
