+++ backend/models/WarehouseBin.js
@@ -9,8 +9,8 @@ const WarehouseBin = sequelize.define("WarehouseBin", {
   capacityUsed: { type: DataTypes.INTEGER, defaultValue: 0 },
   accessCount:  { type: DataTypes.INTEGER, defaultValue: 0 },
   itemCount:    { type: DataTypes.INTEGER, defaultValue: 0 },
-  temperature:  { type: DataTypes.FLOAT, defaultValue: 22.0 },
-  humidity:     { type: DataTypes.FLOAT, defaultValue: 45.0 }
+  temperature:  { type: DataTypes.DECIMAL(15, 2), defaultValue: 22.0 },
+  humidity:     { type: DataTypes.DECIMAL(15, 2), defaultValue: 45.0 }
 });
 
 module.exports = WarehouseBin;
